When creating an EKS cluster, the user who does the creation is assigned special access to be able to connect to the cluster to do the initial setup. This can cause issues with terraform where if another user tries to run the terraform they may not have access to the cluster since they are not the initial user. We were able to work around this in the kubernetes terraform by adding an `exec` block which defined a local command to run to get a token to access the cluster (`aws eks get-token`). This was also not ideal because it depends a lot more on the running user's local k8s setup.
The fix:
We determined that the user-binding-on-cluster-create behaviour also applies to Roles.
This commit has code which adds a role with access to create an EKS cluster, and then uses an AWS provider with an alias to assume that role only while running the EKS module.
Unfortunately we had to move the creation of the new role into the bootstrap because of an order-of-operations issue with trying to assume a role in a provider that was created in the same tf run.